split store_value function into its own class to make tests work
